@@ -2,7 +2,7 @@
   "name": "riot-web",
   "productName": "Revolution",
   "main": "electron_app/src/electron-main.js",
-  "version": "1.3.2",
+  "version": "1.4.2",
   "description": "A feature-rich [Matrix] client",
   "author": "New Vector Ltd. & ponies.im",
   "repository": {
@@ -36,10 +36,10 @@
     "build:compile": "yarn reskindex && babel --source-maps -d lib src",
     "build:bundle": "cross-env NODE_ENV=production webpack -p --progress --bail --mode production",
     "build:bundle:dev": "webpack --progress --bail --mode development",
-    "build:electron": "yarn clean && yarn build && yarn install:electron && build -wml --ia32 --x64",
-    "build:electron:linux": "yarn build && build -l --x64",
-    "build:electron:macos": "yarn build && build -m --x64",
-    "build:electron:windows": "yarn build && build -w --ia32 --x64",
+    "build:electron": "yarn clean && yarn build && yarn install:electron && electron-builder -wml --ia32 --x64",
+    "build:electron:linux": "yarn build && electron-builder -l --x64",
+    "build:electron:macos": "yarn build && electron-builder -m --x64",
+    "build:electron:windows": "yarn build && electron-builder -w --ia32 --x64",
     "build:react-sdk": "node scripts/yarn-sub.js matrix-react-sdk build",
     "build:js-sdk": "node scripts/yarn-sub.js matrix-js-sdk start:init",
     "build": "yarn build:js-sdk && yarn build:react-sdk && yarn reskindex && yarn build:res && yarn build:bundle",
@@ -71,17 +71,17 @@
     "draft-js": "^0.11.0-alpha",
     "extract-text-webpack-plugin": "^4.0.0-beta.0",
     "favico.js": "^0.3.10",
-    "gemini-scrollbar": "github:matrix-org/gemini-scrollbar#b302279",
+    "gemini-scrollbar": "github:matrix-org/gemini-scrollbar#91e1e566",
     "gfm.css": "^1.1.2",
     "highlight.js": "^9.13.1",
     "indexeddbshim": "^3.8.0",
-    "matrix-js-sdk": "2.3.0",
-    "matrix-react-sdk": "1.5.1",
+    "matrix-js-sdk": "github:matrix-org/matrix-js-sdk#develop",
+    "matrix-react-sdk": "github:matrix-org/matrix-react-sdk#develop",
     "modernizr": "^3.6.0",
     "olm": "https://packages.matrix.org/npm/olm/olm-3.1.0.tgz",
-    "prop-types": "^15.6.2",
-    "react": "^15.6.0",
-    "react-dom": "^15.6.0",
+    "prop-types": "^15.7.2",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0",
     "sanitize-html": "^1.19.1",
     "ua-parser-js": "^0.7.19",
     "url": "^0.11.0"
@@ -107,8 +107,8 @@
     "cpx": "^1.3.2",
     "cross-env": "^4.0.0",
     "css-loader": "^2.1.0",
-    "electron-builder": "^20.44.4",
-    "electron-builder-squirrel-windows": "^21.0.6",
+    "electron-builder": "^21.2.0",
+    "electron-builder-squirrel-windows": "^21.2.0",
     "electron-devtools-installer": "^2.2.4",
     "eslint": "^5.8.0",
     "eslint-config-google": "^0.7.1",
@@ -143,8 +143,6 @@
     "postcss-scss": "^1.0.6",
     "postcss-simple-vars": "^4.1.0",
     "postcss-strip-inline-comments": "^0.1.5",
-    "react-addons-perf": "^15.4.0",
-    "react-addons-test-utils": "^15.6.0",
     "rimraf": "^2.4.3",
     "source-map-loader": "^0.2.4",
     "webpack": "^4.23.1",
@@ -153,7 +151,7 @@
   },
   "build": {
     "appId": "im.riot.app",
-    "electronVersion": "5.0.7",
+    "electronVersion": "6.0.11",
     "files": [
       "node_modules/**",
       "src/**"
@@ -173,6 +171,9 @@
       "desktop": {
         "StartupWMClass": "riot"
       }
+    },
+    "deb": {
+      "afterInstall": "electron_app/build/linux/after-install.tpl"
     },
     "mac": {
       "category": "public.app-category.social-networking"
