@@ -33,9 +33,6 @@ import React from 'react';
 // add React and ReactPerf to the global namespace, to make them easier to
 // access via the console
 global.React = React;
-if (process.env.NODE_ENV !== 'production') {
-    global.Perf = require('react-addons-perf');
-}
 
 import './modernizr';
 import ReactDOM from 'react-dom';
@@ -195,11 +192,6 @@ async function loadApp() {
 
     await loadOlm();
 
-    await loadLanguage();
-
-    const fragparts = parseQsFromFragment(window.location);
-    const params = parseQs(window.location);
-
     // set the platform for react sdk
     if (window.ipcRenderer) {
         console.log("Using Electron platform");
@@ -238,6 +230,12 @@ async function loadApp() {
     // XXX: We call this twice, once here and once in MatrixChat as a prop. We call it here to ensure
     // granular settings are loaded correctly and to avoid duplicating the override logic for the theme.
     SdkConfig.put(configJson);
+
+    // Load language after loading config.json so that settingsDefaults.language can be applied
+    await loadLanguage();
+
+    const fragparts = parseQsFromFragment(window.location);
+    const params = parseQs(window.location);
 
     // don't try to redirect to the native apps if we're
     // verifying a 3pid (but after we've loaded the config)
@@ -336,7 +334,8 @@ async function loadApp() {
 
     const acceptInvalidBrowser = window.localStorage && window.localStorage.getItem('mx_accepts_unsupported_browser');
 
-    console.log("Vector starting at "+window.location);
+    const urlWithoutQuery = window.location.protocol + '//' + window.location.host + window.location.pathname;
+    console.log("Vector starting at " + urlWithoutQuery);
     if (configError) {
         window.matrixChat = ReactDOM.render(<div className="error">
             Unable to load config file: please refresh the page to try again.
