@@ -92,9 +92,12 @@ async function initPage() {
     if (isUrl && !isUrl.endsWith('/')) isUrl += '/';
 
     if (hsUrl !== 'https://matrix.org/') {
+        document.getElementById('configure_riot_button').href =
+            "https://riot.im/config/config?hs_url=" + encodeURIComponent(hsUrl) +
+            "&is_url=" + encodeURIComponent(isUrl);
+        document.getElementById('step1_heading').innerHTML= '1: Install the app';
         document.getElementById('step2_container').style.display = 'block';
         document.getElementById('hs_url').innerText = hsUrl;
-        document.getElementById('step_login_header').innerHTML= 'Step 3: Register or Log in';
 
         if (isUrl && isUrl !== "https://vector.im/") {
             document.getElementById('default_is').style.display = 'none';
