@@ -34,7 +34,7 @@ const RoomView = sdk.getComponent('structures.RoomView');
 
 const React = require('react');
 const ReactDOM = require('react-dom');
-const ReactTestUtils = require('react-addons-test-utils');
+const ReactTestUtils = require('react-dom/test-utils');
 const expect = require('expect');
 import Promise from 'bluebird';
 import {makeType} from "matrix-react-sdk/lib/utils/TypeUtils";
