@@ -21,6 +21,8 @@ module.exports = {
         "theme-dark": "./node_modules/matrix-react-sdk/res/themes/dark/css/dark.scss",
         //"theme-status": "./res/themes/status/css/status.scss",
         //"theme-blueshine": "./res/themes/blueshine/css/blueshine.scss",
+        "theme-light-custom": "./node_modules/matrix-react-sdk/res/themes/light-custom/css/light-custom.scss",
+        "theme-dark-custom": "./node_modules/matrix-react-sdk/res/themes/dark-custom/css/dark-custom.scss",
     },
     module: {
         rules: [
@@ -155,7 +157,6 @@ module.exports = {
             // using `npm link` / `yarn link`.
             "react": path.resolve('./node_modules/react'),
             "react-dom": path.resolve('./node_modules/react-dom'),
-            "react-addons-perf": path.resolve('./node_modules/react-addons-perf'),
 
             // same goes for js-sdk
             "matrix-js-sdk": path.resolve('./node_modules/matrix-js-sdk'),
